requirements.txt
# pip3 install -r requirements.txt

pandas==1.0.1
numpy~=1.18.1
scipy~=1.4.1
seaborn==0.9.0
matplotlib==3.1.0
scikit-learn==0.21.2
tensorflow-gpu==1.15.0
keras==2.2.4
opencv-python==4.2.0.34
scikit-image==0.15.0
imutils==0.5.3
Flask~=1.1.1
mysql-connector-python~=8.0.21
PyYAML~=5.3